move clusterbuildstrategy
